requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.5
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==4.2.7
django-allauth==0.58.2
django-bootstrap-v5==1.0.11
django-cache-url==3.4.4
django-crispy-forms==2.1
elevenlabs==0.2.26
environs==9.5.0
executing==2.0.1
ffmpeg==1.4
frozenlist==1.4.0
gunicorn==20.1.0
idna==3.4
ipython==8.18.0
jedi==0.19.1
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.6
packaging==23.2
parso==0.8.3
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg==3.1.8
psycopg-binary==3.1.8
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
websockets==12.0
whitenoise==6.4.0
yarl==1.9.3
